chore: [refactoring][KanuKim97]: navigation 정리
1. 네비게이션 타입 분리
  - 기존 StackTypes.ts 분리
  - 이에 따른 import 변경
2. Header 설정 관련 option Helper 함수 추가
3. Tab navigator ownership 정리
 - createBottomTabNavigator()는 TabsNavigator.tsx에서만 생성한다.
 - BottomNavigationBar는 navigator를 생성하지 않고, tab icon/label/screen option helper만 제공하도록 바꾼다.
 -"createBottomTabNavigator" src 결과가 TabsNavigator.tsx 1곳만 남게 한다.
4. Shared 정리 준비
   - calendar/personal, team, header는 이동 대상 목록으로 확정
  - formatGetData.ts의 잘못된 calenderType import는 TimeFrame 기준으로 수정한다.
