+++ src/navigation/RootNavigator.tsx
@@ -5,7 +5,11 @@ import SplashScreen from '../presentation/splash/SplashScreen'
 import TabsNavigator from './TabsNavigator'
 import OnBoardingScheduleNavigator from './OnboardingScheduleNavigator'
 import LoginNavigator from './LoginNavigator'
-import { RootStackParamList } from './types/StackTypes'
+import { RootStackParamList } from './types/RootStackParamList'
+import {
+  emptyHeaderOptions,
+  makeBackHeaderOptions,
+} from './options/stackHeaderOptions'
 
 // Migrate from MainNavigator.tsx
 import TodoScreen from '../presentation/note/screens/TodoScreen'
@@ -30,9 +34,6 @@ import OnboardingMethodScreen from '../presentation/onboarding/screens/Onboardin
 // Migrate from AutoAlarmNavigator.tsx
 import CreateAutoAlarmScreen from '../presentation/alarm/screen/CreateAutoAlarmScreen'
 import EditAutoAlarmScreen from '../presentation/alarm/screen/EditAutoAlarmScreen'
-import CenterAlignedTopAppBar from '../shared/components/appbar/CenterAlignedTopAppBar'
-import TopAppBarBackButton from '../shared/components/button/TopAppBarBackButton'
-import GlobalText from '../shared/components/text/GlobalText'
 
 // Onboarding Navigation Error Screen
 const RootStack = createNativeStackNavigator<RootStackParamList>()
@@ -47,17 +48,7 @@ const RootNavigator = () => {
           // OCR / NEW / DIRECT 선택 화면
           name="OnboardingMethodScreen"
           component={OnboardingMethodScreen}
-          options={() => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={null}
-                title={null}
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={emptyHeaderOptions}
         />
         <RootStack.Screen
           name="OnboardingSchedules"
@@ -70,61 +61,18 @@ const RootNavigator = () => {
         <RootStack.Screen
           name="Todo"
           component={TodoScreen}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    할 일
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={makeBackHeaderOptions('할 일')}
         />
         <RootStack.Screen
           name="Memo"
           component={MemoScreen}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    메모
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={makeBackHeaderOptions('메모')}
         />
         <RootStack.Screen
           name="AddMemo"
           component={AddMemoScreen}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={null}
-                applySafeArea={true}
-                backgroundColor="bg-surface-white"
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
+          options={makeBackHeaderOptions(null, {
+            backgroundColor: 'bg-surface-white',
           })}
         />
 
@@ -139,156 +87,44 @@ const RootNavigator = () => {
         <RootStack.Screen
           name="EditProfileScreen"
           component={EditProfileScreen}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    프로필 수정
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={makeBackHeaderOptions('프로필 수정')}
         />
 
         <RootStack.Screen
           name="FeedbackScreen"
           component={FeedBackScreen}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    평가 및 피드백
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={makeBackHeaderOptions('평가 및 피드백')}
         />
 
         <RootStack.Screen
           name="TermsWebViewScreen"
           component={TermsWebViewScreen}
-          options={({ navigation, route }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    {route.params.title}
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={({ navigation, route }) =>
+            makeBackHeaderOptions(route.params.title)({ navigation })
+          }
         />
 
         <RootStack.Screen
           name="WithdrawBeforeScreen"
           component={WithdrawBeforeScreen}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    회원 탈퇴
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={makeBackHeaderOptions('회원 탈퇴')}
         />
         <RootStack.Screen
           name="WithdrawScreen"
           component={WithdrawScreen}
-          options={({ navigation }) => {
-            return {
-              header: () => (
-                <CenterAlignedTopAppBar
-                  navigationIcon={
-                    <TopAppBarBackButton onPress={navigation.goBack} />
-                  }
-                  title={
-                    <GlobalText className="font-pretSemiBold text-heading-xs">
-                      회원 탈퇴
-                    </GlobalText>
-                  }
-                  applySafeArea={true}
-                />
-              ),
-              headerShown: true,
-              headerShadowVisible: false,
-            }
-          }}
+          options={makeBackHeaderOptions('회원 탈퇴')}
         />
 
         {/* Migrate from AutoAlarmNavigator.tsx */}
         <RootStack.Screen
           name="CreateAlarm"
           component={CreateAutoAlarmScreen}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    알람 생성
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={makeBackHeaderOptions('알람 생성')}
         />
         <RootStack.Screen
           name="AutoAlarm"
           component={AutoAlarm}
-          options={({ navigation }) => ({
-            header: () => (
-              <CenterAlignedTopAppBar
-                navigationIcon={
-                  <TopAppBarBackButton onPress={navigation.goBack} />
-                }
-                title={
-                  <GlobalText className="font-pretSemiBold text-heading-xs">
-                    자동 알람
-                  </GlobalText>
-                }
-                applySafeArea={true}
-              />
-            ),
-            headerShown: true,
-            headerShadowVisible: false,
-          })}
+          options={makeBackHeaderOptions('자동 알람')}
         />
         <RootStack.Screen
           name="EditAutoAlarm"
