+++ src/navigation/types/LoginStackParamList.ts
@@ -1,0 +1,5 @@
+export type LoginStackParamList = {
+  Login: undefined
+  KakaoWebView: undefined
+  TermsWebViewScreen: { title: string; url: string }
+}
