+++ src/navigation/types/TabParamList.ts
@@ -1,0 +1,6 @@
+export type TabParamList = {
+  Home: undefined
+  Calendar: { selectedDate?: string; isTeamView?: boolean } | undefined
+  MyInformation: undefined
+  AutoAlarm: undefined
+}
