+++ src/presentation/alarm/screen/AutoAlarmScreen.tsx
@@ -13,7 +13,7 @@ import AlarmPlus from '../../../assets/icons/alarm_plus_24.svg'
 import SettingsIcon from '../../../assets/icons/alarm_three-dot_24.svg'
 import { Portal } from '@gorhom/portal'
 import { useNavigation } from '@react-navigation/native'
-import { rootNavigation } from '../../../navigation/types/StackTypes'
+import { rootNavigation } from '../../../navigation/types/NavigationProps'
 import StartAlignedTopAppBar from '../../../shared/components/appbar/StartAlignedTopAppBar'
 import GlobalText from '../../../shared/components/text/GlobalText'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
