+++ src/presentation/calendar/component/PlusEdit.tsx
@@ -9,7 +9,7 @@ import CancelIcon from '../../../assets/icons/w-cancel.svg'
 import CameraIcon from '../../../assets/icons/pr-cam.svg'
 import PencilIcon from '../../../assets/icons/pr-pencil.svg'
 import { useNavigation } from '@react-navigation/native'
-import { rootNavigation } from '../../../navigation/types/StackTypes'
+import { rootNavigation } from '../../../navigation/types/NavigationProps'
 import { Animated } from 'react-native'
 import { useScheduleInfoStore } from '../../../store/useScheduleInfoStore'
 import { useOnboardingStore } from '../../../store/useOnboardingStore'
