+++ src/presentation/info/screen/WithdrawBeforeScreen.tsx
@@ -10,7 +10,7 @@ import { SafeAreaView } from 'react-native-safe-area-context'
 import GlobalText from '../../../shared/components/text/GlobalText'
 import CheckBoxMenuItem from '../components/CheckBoxMenuItem'
 import CheckBoxWithTextInput from '../components/CheckBoxWithTextInput'
-import { rootNavigation } from '../../../navigation/types/StackTypes'
+import { rootNavigation } from '../../../navigation/types/NavigationProps'
 import { useNavigation } from '@react-navigation/native'
 import { useState } from 'react'
 import EmphasizedButton from '../../../shared/components/button/Button'
