+++ src/presentation/info/screen/WithdrawScreen.tsx
@@ -1,7 +1,7 @@
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { ActivityIndicator, Alert, View } from 'react-native'
 import GlobalText from '../../../shared/components/text/GlobalText'
-import { rootNavigation } from '../../../navigation/types/StackTypes'
+import { rootNavigation } from '../../../navigation/types/NavigationProps'
 import { CommonActions, useNavigation } from '@react-navigation/native'
 import { useState } from 'react'
 import { memberRepository } from '../../../infrastructure/di/Dependencies'
