+++ src/presentation/onboarding/screens/CompleteScheduleScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { View } from 'react-native'
 import { useNavigation } from '@react-navigation/native'
-import { RootStackParamList } from '../../../navigation/types/StackTypes'
+import { RootStackParamList } from '../../../navigation/types/RootStackParamList'
 import { NativeStackNavigationProp } from '@react-navigation/native-stack'
 import HeadLineText from '../../../shared/components/text/HeadLineText'
 import { SafeAreaView } from 'react-native-safe-area-context'
