+++ src/presentation/onboarding/screens/ocr/SelectPhotoOCRScreen.tsx
@@ -16,7 +16,7 @@ import RegMethod from '../../component/RegMethod'
 import TakePicture from '../../../../assets/icons/ic_camera_32.svg'
 import OpenGallery from '../../../../assets/icons/ic_gallery_32.svg'
 import { RouteProp, useNavigation, useRoute } from '@react-navigation/native'
-import { OnboardingStackParamList } from '../../../../navigation/types/StackTypes'
+import { OnboardingStackParamList } from '../../../../navigation/types/OnboardingStackParamList'
 import ProgressModal from '../../../../shared/components/progress/ProgressModal'
 import goNextOnboadingScreen from '../../flow/goNextOnboardingScreen'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
